fix liquibase

+++ src/main/resources/db/changelog/changes/010-create-snack-image.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS snacks_images(
-    id INT  PRIMARY KEY AUTO_INCREMENT,
+    id BIGINT  PRIMARY KEY AUTO_INCREMENT,
     image LONGBLOB NOT NULL,
     name VARCHAR(500) NOT NULL,
-    snack_id INT,
+    snack_id BIGINT,
     CONSTRAINT fk_snacks FOREIGN KEY (snack_id) REFERENCES snacks(id) ON DELETE CASCADE
 
 );
