+++ src/main/resources/db/changelog/changes/017-create-product-rating.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS product_bundle_rating (
     id BIGINT AUTO_INCREMENT PRIMARY KEY,
     rating DOUBLE NOT NULL,
-    product_id INT NOT NULL,
+    product_id BIGINT NOT NULL,
     CONSTRAINT fk_product_rating FOREIGN KEY (product_id) REFERENCES products_bundle(id) ON DELETE CASCADE
 );
