+++ src/main/resources/db/changelog/changes/021-create-snack-offer.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS snack_special_offer_product (
-    snack_id INT,
-    special_offer_product_id INT,
+    snack_id BIGINT,
+    special_offer_product_id BIGINT,
     PRIMARY KEY (snack_id, special_offer_product_id),
     FOREIGN KEY (snack_id) REFERENCES snacks(id),
     FOREIGN KEY (special_offer_product_id) REFERENCES special_offer_product(id)
