+++ src/main/resources/db/changelog/changes/030-create-beer-options.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS beer_options (
-      id INT AUTO_INCREMENT PRIMARY KEY,
-      beer_id INT NOT NULL,
+      id BIGINT AUTO_INCREMENT PRIMARY KEY,
+      beer_id BIGINT NOT NULL,
       volume DECIMAL(7,2) NOT NULL,
       quantity INT NOT NULL,
       price DECIMAL(7,2) NOT NULL,
