+++ src/main/resources/db/changelog/changes/054-create-order-item.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS order_items (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    order_id INT NOT NULL,
-    item_id INT NOT NULL,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    order_id BIGINT NOT NULL,
+    item_id BIGINT NOT NULL,
     item_type VARCHAR(50) NOT NULL,
     quantity INT NOT NULL,
     price DECIMAL(7,2) NOT NULL,
